requirements.txt
apturl==0.5.2
attrs==24.1.0
awscli==1.22.34
bcrypt==3.2.0
blinker==1.7.0
boto3==1.34.79
botocore==1.34.79
Brlapi==0.8.3
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
duplicity==0.8.21
exceptiongroup==1.2.2
fasteners==0.14.1
Flask==3.0.2
future==0.18.2
gunicorn==23.0.0
gyp==0.1
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyvips==2.2.3
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
roman==3.3
rsa==4.8
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.1
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
systemd-python==234
threadpoolctl==3.3.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
